# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Russian Electric Trains (Elektrichki)
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Station {
  code: string;
  title: string;
  popular_title?: string;
  short_title?: string;
  lat?: number | null;
  lng?: number | null;
  region?: string | null;
  direction?: string | null;
}

export interface StationsResponse {
  stations: Station[];
}

export interface TrainSegment {
  uid: string;
  number: string;
  title: string;
  departure: string;
  arrival: string;
  duration: number;
  from_station: Station;
  to_station: Station;
  stops?: string | null;
  price_min?: number | null;
  price_max?: number | null;
  ticket_type?: string | null;
  transport_subtype?: string | null;
  thread_express_type?: string | null;
}

export interface ScheduleResponse {
  date: string;
  from: Station;
  to: Station;
  segments: TrainSegment[];
  interval_segments: TrainSegment[];
}

export type TicketStatus = (typeof TicketStatus)[keyof typeof TicketStatus];

export const TicketStatus = {
  active: "active",
  used: "used",
  cancelled: "cancelled",
} as const;

export interface Ticket {
  id: string;
  segment: TrainSegment;
  date: string;
  saved_at: string;
  status: TicketStatus;
}

export interface TicketsResponse {
  tickets: Ticket[];
}

export interface SaveTicketRequest {
  segment: TrainSegment;
  date: string;
}

export type SearchStationsParams = {
  q: string;
};

export type GetScheduleParams = {
  /**
   * Yandex station code (from)
   */
  from: string;
  /**
   * Yandex station code (to)
   */
  to: string;
  /**
   * Date in YYYY-MM-DD format (defaults to today)
   */
  date?: string;
};