# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification for Russian Electric Trains (Elektrichki)
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tickets:
    get:
      operationId: getTickets
      tags: [tickets]
      summary: Get user saved tickets
      responses:
        "200":
          description: Saved tickets
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketsResponse"

    post:
      operationId: saveTicket
      tags: [tickets]
      summary: Save a ticket
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveTicketRequest"
      responses:
        "201":
          description: Ticket saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
  /tickets/{id}:
    delete:
      operationId: deleteTicket
      tags: [tickets]
      summary: Delete a ticket
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Ticket deleted
components:
  schemas:
...
    Station:
      type: object
      properties:
        code:
          type: string
        title:
          type: string
        popular_title:
          type: string
        short_title:
          type: string
        lat:
          type: number
          nullable: true
        lng:
          type: number
          nullable: true
        region:
          type: string
          nullable: true
        direction:
          type: string
          nullable: true
      required:
        - code
        - title
...
    TrainSegment:
      type: object
      properties:
        uid:
          type: string
        number:
          type: string
        title:
          type: string
        departure:
          type: string
        arrival:
          type: string
        duration:
          type: number
        from_station:
          $ref: "#/components/schemas/Station"
        to_station:
          $ref: "#/components/schemas/Station"
        stops:
          type: string
          nullable: true
        price_min:
          type: number
          nullable: true
        price_max:
          type: number
          nullable: true
        ticket_type:
          type: string
          nullable: true
        transport_subtype:
          type: string
          nullable: true
        thread_express_type:
          type: string
          nullable: true
      required:
        - uid
        - number
        - title
        - departure
        - arrival
        - duration
        - from_station
        - to_station
...
    Ticket:
      type: object
      properties:
        id:
          type: string
        segment:
          $ref: "#/components/schemas/TrainSegment"
        date:
          type: string
        saved_at:
          type: string
        status:
          type: string
          enum: [active, used, cancelled]
      required:
        - id
        - segment
        - date
        - saved_at
        - status
    TicketsResponse:
      type: object
      properties:
        tickets:
          type: array
          items:
            $ref: "#/components/schemas/Ticket"
      required:
        - tickets
    SaveTicketRequest:
      type: object
      properties:
        segment:
          $ref: "#/components/schemas/TrainSegment"
        date:
          type: string
      required:
        - segment
        - date