# lib/api-zod/src/generated/types/saveTicketRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Russian Electric Trains (Elektrichki)
 * OpenAPI spec version: 0.1.0
 */
import type { TrainSegment } from "./trainSegment";

export interface SaveTicketRequest {
  segment: TrainSegment;
  date: string;
}