# lib/api-zod/src/generated/types/station.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Russian Electric Trains (Elektrichki)
 * OpenAPI spec version: 0.1.0
 */

export interface Station {
  code: string;
  title: string;
  popular_title?: string;
  short_title?: string;
  lat?: number | null;
  lng?: number | null;
  region?: string | null;
  direction?: string | null;
}