# lib/api-zod/src/generated/types/ticket.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Russian Electric Trains (Elektrichki)
 * OpenAPI spec version: 0.1.0
 */
import type { TicketStatus } from "./ticketStatus";
import type { TrainSegment } from "./trainSegment";

export interface Ticket {
  id: string;
  segment: TrainSegment;
  date: string;
  saved_at: string;
  status: TicketStatus;
}