# lib/api-zod/src/generated/types/ticketsResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Russian Electric Trains (Elektrichki)
 * OpenAPI spec version: 0.1.0
 */
import type { Ticket } from "./ticket";

export interface TicketsResponse {
  tickets: Ticket[];
}