# lib/api-zod/src/generated/types/trainSegment.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Russian Electric Trains (Elektrichki)
 * OpenAPI spec version: 0.1.0
 */
import type { Station } from "./station";

export interface TrainSegment {
  uid: string;
  number: string;
  title: string;
  departure: string;
  arrival: string;
  duration: number;
  from_station: Station;
  to_station: Station;
  stops?: string | null;
  price_min?: number | null;
  price_max?: number | null;
  ticket_type?: string | null;
  transport_subtype?: string | null;
  thread_express_type?: string | null;
}